money_bot/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
crontab==0.23.0
idna==3.3
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
